[rs] Refactor top-level swf parser
