+++ rs/fuzz/Cargo.toml
@@ -1,4 +1,3 @@
-
 [package]
 name = "swf-parser-fuzz"
 version = "0.0.1"
