+++ rs/fuzz/fuzz_targets/movie.rs
@@ -2,9 +2,6 @@
 #[macro_use] extern crate libfuzzer_sys;
 extern crate swf_parser;
 
-//use swf_parser;
-//use swf_tree;
-
 fuzz_target!(|data: &[u8]| {
-   let _ = swf_parser::parsers::movie::parse_movie(&data[..]);
+   let _ = swf_parser::complete::parse_swf(data);
 });
